clases en archivos individuales por entidad en la carpeta Entidades
