+++ lib_dominio/Entidades/Snacks.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace lib_dominio.Entidades
+{
+    public class Snacks
+    {
+        [Key]
+        public int Id { get; set; }
+        public string? Nombre { get; set; }
+        public decimal Precio { get; set; }
+        public int Stock { get; set; }
+    }
+}
